// src/Models/SalesReportTypes.ts
export interface SaleItem {
    date: string;
    paid_to: string;
    item_type: string;
    item_name: string;
    artist: string;
    currency: string;
    item_price: number;
    quantity: number;
    discount_code: string;
    sub_total: number;
    shipping: number;
    ship_from_country_name: string;
    transaction_fee: number;
    fee_type: string;
    item_total: number;
    amount_you_received: number;
    bandcamp_transaction_id: string;
    paypal_transaction_id: string;
    net_amount: number;
    package: string;
    option: string;
    item_url: string;
    catalog_number: string;
    upc: string;
    isrc: string;
    buyer_name: string;
    buyer_email: string;
    buyer_phone: string;
    buyer_note: string;
    ship_to_name: string;
    ship_to_street: string;
    ship_to_street_2: string;
    ship_to_city: string;
    ship_to_state: string;
    ship_to_zip: string;
    ship_to_country: string;
    ship_to_country_code: string;
    ship_date: string;
    ship_notes: string;
    country: string;
    country_code: string;
    region_or_state: string;
    city: string;
    referer: string;
    referer_url: string;
    sku: string;
    seller_tax: number;
    marketplace_tax: number;
}